Renaming redundantly named file.
