+++ src/lib/engine/deal-generator/index.ts
